deny `package-preview-2023-07` tag for azure_mgmt_alertsmanagement due to duplicate struct names
