+++ services/mgmt/alertsmanagement/Cargo.toml
@@ -33,12 +33,12 @@ env_logger = "0.10"
 all-features = true
 
 [features]
-default = ["package-preview-2023-07", "enable_reqwest"]
+default = ["package-preview-2023-05", "enable_reqwest"]
 enable_reqwest = ["azure_core/enable_reqwest"]
 enable_reqwest_rustls = ["azure_core/enable_reqwest_rustls"]
 no-default-tag = []
-"package-preview-2023-07" = []
 "package-preview-2023-05" = []
 "package-preview-2023-04" = []
 "package-preview-2023-01" = []
 "package-preview-2021-08" = []
+"package-preview-2021-07" = []
