+++ services/mgmt/alertsmanagement/README.md
@@ -6,12 +6,12 @@ https://github.com/Azure/azure-rest-api-specs/blob/main/specification/alertsmana
 
 To get started with these generated service crates, see the [examples](https://github.com/Azure/azure-sdk-for-rust/blob/main/services/README.md#examples).
 
-The default tag is `package-preview-2023-07`.
+The default tag is `package-preview-2023-05`.
 
 The following [tags](https://github.com/Azure/azure-sdk-for-rust/blob/main/services/tags.md) are available:
 
-- `package-preview-2023-07` has 9 operations from 1 API versions: `2023-07-12-preview`. Use crate feature `package-preview-2023-07` to enable. The operations will be in the `package_preview_2023_07` module.
 - `package-preview-2023-05` has 6 operations from 1 API versions: `2023-05-01-preview`. Use crate feature `package-preview-2023-05` to enable. The operations will be in the `package_preview_2023_05` module.
 - `package-preview-2023-04` has 6 operations from 1 API versions: `2023-04-01-preview`. Use crate feature `package-preview-2023-04` to enable. The operations will be in the `package_preview_2023_04` module.
 - `package-preview-2023-01` has 2 operations from 1 API versions: `2023-01-01-preview`. Use crate feature `package-preview-2023-01` to enable. The operations will be in the `package_preview_2023_01` module.
 - `package-preview-2021-08` has 17 operations from 2 API versions: `2019-05-05-preview`, `2021-08-08-preview`. Use crate feature `package-preview-2021-08` to enable. The operations will be in the `package_preview_2021_08` module.
+- `package-preview-2021-07` has 6 operations from 1 API versions: `2021-07-22-preview`. Use crate feature `package-preview-2021-07` to enable. The operations will be in the `package_preview_2021_07` module.
