+++ services/mgmt/alertsmanagement/autorust.toml
@@ -1,0 +1,5 @@
+[tags]
+deny = [
+    # error[E0428]: the name `ErrorResponse` is defined multiple times
+    "package-preview-2023-07"
+]
