+++ services/mgmt/alertsmanagement/src/lib.rs
@@ -3,10 +3,6 @@
 #![allow(clippy::ptr_arg)]
 #![allow(clippy::large_enum_variant)]
 #![allow(clippy::derive_partial_eq_without_eq)]
-#[cfg(feature = "package-preview-2023-07")]
-pub mod package_preview_2023_07;
-#[cfg(all(feature = "package-preview-2023-07", not(feature = "no-default-tag")))]
-pub use package_preview_2023_07::*;
 #[cfg(feature = "package-preview-2023-05")]
 pub mod package_preview_2023_05;
 #[cfg(all(feature = "package-preview-2023-05", not(feature = "no-default-tag")))]
@@ -23,3 +19,7 @@ pub use package_preview_2023_01::*;
 pub mod package_preview_2021_08;
 #[cfg(all(feature = "package-preview-2021-08", not(feature = "no-default-tag")))]
 pub use package_preview_2021_08::*;
+#[cfg(feature = "package-preview-2021-07")]
+pub mod package_preview_2021_07;
+#[cfg(all(feature = "package-preview-2021-07", not(feature = "no-default-tag")))]
+pub use package_preview_2021_07::*;
